Split forms.py into several modules
