+++ dmutils/forms/fields.py
@@ -1,0 +1,30 @@
+from itertools import chain
+
+from wtforms import StringField
+from wtforms.validators import Length
+
+from .filters import strip_whitespace
+from .validators import EmailValidator
+
+
+class StripWhitespaceStringField(StringField):
+    def __init__(self, label=None, **kwargs):
+        kwargs['filters'] = tuple(chain(
+            kwargs.get('filters', ()),
+            (
+                strip_whitespace,
+            ),
+        ))
+        super(StringField, self).__init__(label, **kwargs)
+
+
+class EmailField(StripWhitespaceStringField):
+    def __init__(self, label=None, **kwargs):
+        kwargs["validators"] = tuple(chain(
+            kwargs.pop("validators", ()),
+            (
+                EmailValidator(),
+                Length(max=511, message="Please enter an email address under 512 characters."),
+            ),
+        ))
+        super(EmailField, self).__init__(label, **kwargs)
